Plot HIP errors by field

Create a bar plot of proportion of error per field. The plot defaults to all 49 states, but location can be specified.

Usage

errorPlotFields(proofed_data, loc = "all", year)

Arguments

proofed_data

The object created after error flagging data with proof or correct

loc

The location that errors should be plotted for. Acceptable values include:

"all" - all states

a two-letter abbreviation for a US state; one of:

AL, AK, AZ, AR, CA, CO, CT, DE, FL, GA, ID, IL, IN, IA, KS, KY, LA, ME, MD, MA, MI, MN, MS, MO, MT, NE, NV, NH, NJ, NM, NY, NC, ND, OH, OK, OR, PA, RI, SC, SD, TN, TX, UT, VT, VA, WA, WV, WI, WY

year

The year in which the HIP data were collected.

See also

Author

Abby Walter, abby_walter@fws.gov