Table of errors

Create a tibble of existing errors in the data, customizing the output with value specifications.

Usage

errorTable(proofed_data, loc = "all", field = "all")

Arguments

proofed_data

The object created after error flagging data with proof or correct

loc

Which location the error data should be tabulated by. Acceptable values include:

a two-letter abbreviation for a US state; one of:

AL, AK, AZ, AR, CA, CO, CT, DE, FL, GA, ID, IL, IN, IA, KS, KY, LA, ME, MD, MA, MI, MN, MS, MO, MT, NE, NV, NH, NJ, NM, NY, NC, ND, OH, OK, OR, PA, RI, SC, SD, TN, TX, UT, VT, VA, WA, WV, WI, WY

"all" - all states

"none" - table will not include location in its output

field

Field the error data should be tabulated by. Acceptable values include:

If loc = "none", field must be "all". Otherwise, choose one of:

title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe, rails_gallinules, cranes, band_tailed_pigeon, brant, seaducks, registration_yr, email

"all" - all fields

"none" - table will not include field in its output

See also

Author

Abby Walter, abby_walter@fws.gov