Get bad last name values

The internal getBadLastName function is used inside of proof and qualityMessages to filter to lastname values that are not expected. Used by qLastName.

Usage

getBadLastName(data)

Arguments

data

Harvest Information Program registration data

See also

Author

Abby Walter, abby_walter@fws.gov